requirements.txt
python-binance==1.0.16
setuptools==45
Flask==2.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.1
gunicorn==20.1.0
flask_healthz==0.0.3